+++ backend/meshive/models/__init__.py
@@ -1,32 +1,4 @@
-from meshive.models.library_source import LibrarySource
-from meshive.models.creator import CreatorLink
-from meshive.models.favorite import FavoriteList, FavoriteListItem
-from meshive.models.session import UserSession
-from meshive.models.tag import FolderTagRule, ModelTag, Tag
 from meshive.models.backup import BackupRun, BackupSchedule
-from meshive.models.user import User
-from meshive.models.user_token import UserActionToken
-
-__all__ = [
-    "Archive",
-    "ArchiveEntry",
-    "CreatorLink",
-    "FavoriteList",
-    "FavoriteListItem",
-    "LibraryModel",
-    "LibrarySource",
-    "ModelImage",
-    "ScanIssue",
-    "ScanRun",
-    "User",
-    "UserSession",
-    "UserActionToken",
-    "Tag",
-    "ModelTag",
-    "FolderTagRule",
-    "BackupRun",
-    "BackupSchedule",
-]
 from meshive.models.catalog import (
     Archive,
     ArchiveEntry,
@@ -35,3 +7,33 @@ from meshive.models.catalog import (
     ScanIssue,
     ScanRun,
 )
+from meshive.models.creator import CreatorLink
+from meshive.models.favorite import FavoriteList, FavoriteListItem
+from meshive.models.library_source import LibrarySource
+from meshive.models.metadata import MetadataArtwork
+from meshive.models.session import UserSession
+from meshive.models.tag import FolderTagRule, ModelTag, Tag
+from meshive.models.user import User
+from meshive.models.user_token import UserActionToken
+
+__all__ = [
+    "Archive",
+    "ArchiveEntry",
+    "BackupRun",
+    "BackupSchedule",
+    "CreatorLink",
+    "FavoriteList",
+    "FavoriteListItem",
+    "FolderTagRule",
+    "LibraryModel",
+    "LibrarySource",
+    "MetadataArtwork",
+    "ModelImage",
+    "ModelTag",
+    "ScanIssue",
+    "ScanRun",
+    "Tag",
+    "User",
+    "UserActionToken",
+    "UserSession",
+]